신청자 관리 / 참여자 리스트 개선 (#1067)
* docs: 프로젝트 setting 설정 추가

* refactor: 신청자/모임장 management 파일 분리

* feat: 신청자/모임장 리스트 넘버링

* refactor: util 함수 분리

* refactor: API 성공/실패 처리 로직을 컴포넌트에서 hook으로 이동

* refactor: 신청 상태별 버튼 로직 설정 객체로 중앙화하여 중복 코드 제거

* refactor: 신청 상태별 버튼 로직 설정 객체로 중앙화하여 중복 코드 제거

* refactor: alt 속성 추가
